requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
graphviz>=0.20.1
langchain_tavily>=0.1.0
tavily-python>=0.3.0
langsmith>=0.1.0
pathlib2>=2.3.7; python_version < "3.4"
